Fixed the load method on the base image divider class
Loading isn't supported in image division but the return type had to be fixed.

+++ src/formats/divide/ImageDividerConverter.cs
@@ -59,7 +59,7 @@ namespace Formats.Divide
 
 		protected ImageDividerConverter(string name) : base(name) { }
 		public override string FilterString { get { return "*.*"; } }
-		public override byte[][] Load(Hashtable input) 
+		public override int[][] Load(Hashtable input) 
 		{
 			throw new NotImplementedException("Load is not supported in this filter");
 		}
